refactor(cli): extract randomNames to its own file
